backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
pydantic>=2.10.0
pydantic-settings>=2.5.0
SQLAlchemy>=2.0.27
aiosqlite>=0.19.0
python-dotenv>=1.0.1
loguru>=0.7.2

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
openpyxl>=3.1.2
